ensure power off works after istep boot
istep is used to boot a eBMC based system outside of its normal systemd
targets and services. This is fine for the boot but we then rely on the
normal eBMC services to power off the system. But because the normal
eBMC services were not used to power on, the corresponding services
needed to power off are not in synch. This new obmcutil command was
added to allow users to get their eBMC services back in synch when using
an external tool to boot the system.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>